Split integration test into multiple files
