+++ tests/upload_files.rs
@@ -1,0 +1,51 @@
+mod helpers;
+use helpers::*;
+
+#[rstest]
+fn uploading_files_works(tmpdir: TempDir, port: u16) -> Result<(), Error> {
+    let test_file_name = "uploaded test file.txt";
+
+    let mut child = Command::cargo_bin("miniserve")?
+        .arg(tmpdir.path())
+        .arg("-p")
+        .arg(port.to_string())
+        .arg("-u")
+        .stdout(Stdio::null())
+        .spawn()?;
+
+    sleep(Duration::from_secs(1));
+
+    // Before uploading, check whether the uploaded file does not yet exist.
+    let body = reqwest::get(format!("http://localhost:{}", port).as_str())?.error_for_status()?;
+    let parsed = Document::from_read(body)?;
+    assert!(parsed.find(Text).all(|x| x.text() != test_file_name));
+
+    // Perform the actual upload.
+    let upload_action = parsed
+        .find(Attr("id", "file_submit"))
+        .next()
+        .expect("Couldn't find element with id=file_submit")
+        .attr("action")
+        .expect("Upload form doesn't have action attribute");
+    let form = multipart::Form::new();
+    let part = multipart::Part::text("this should be uploaded")
+        .file_name(test_file_name)
+        .mime_str("text/plain")?;
+    let form = form.part("file_to_upload", part);
+
+    let client = reqwest::Client::new();
+    client
+        .post(format!("http://localhost:{}{}", port, upload_action).as_str())
+        .multipart(form)
+        .send()?
+        .error_for_status()?;
+
+    // After uploading, check whether the uploaded file is now getting listed.
+    let body = reqwest::get(format!("http://localhost:{}", port).as_str())?;
+    let parsed = Document::from_read(body)?;
+    assert!(parsed.find(Text).any(|x| x.text() == test_file_name));
+
+    child.kill()?;
+
+    Ok(())
+}
